requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.18
django-allauth==0.50.0
django-cors-headers==3.12.0
django-countries==7.5.1
django-crispy-forms==1.14.0
django-filter>=21.1
django-formtools==2.3
django-redis
django-seed==0.3.1
django-select2
django-tables2==2.4.1
django-debug-toolbar==3.4.0
django-rest-auth==0.9.5
dj-rest-auth==2.2.4
djangorestframework-simplejwt==5.1.0
Faker==17.3.0
django-mathfilters==1.0.0
gprof2dot==2022.7.29
idna==3.4
lxml==4.9.2
oauthlib
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
toposort==1.10
typing_extensions==4.4.0
urllib3==1.26.14
webencodings==0.5.1
boto3

django-import-export==3.3.1
